configuração criação de controllers e rotas

+++ src/modules/products/typeorm/entities/products_entity.ts
@@ -12,7 +12,7 @@ class ProductsEntity {
   @Column('decimal')
   price: number;
 
-  @Column('int')
+  @Column('decimal')
   quantity: number;
 
   @CreateDateColumn()
